Fixed some inconsistencies

+++ src/main/java/com/sublinks/sublinksapi/api/lemmy/v3/comment/models/ListCommentLikes.java
@@ -4,7 +4,7 @@ import lombok.Builder;
 
 @Builder
 public record ListCommentLikes(
-    Long comment_id,
+    Integer comment_id,
     Integer limit,
     Integer page
 ) {
